refactor: global app cleanup

+++ server/src/controllers/group.controller.js
@@ -1,11 +1,16 @@
 
 import * as groupService from "../services/group.service.js"
 
+const sendServiceResult = (res, result, successStatus = 200) => {
+  const status = Number(result?.statusCode) || successStatus;
+  res.status(status).json(result);
+};
+
 const getGroup = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const channel = await groupService.getGroup(id);
-    res.json(channel);
+    const channel = await groupService.getGroup(id, req.user.id);
+    sendServiceResult(res, channel);
   } catch (error) {
     next(error);
   }
@@ -15,7 +20,7 @@ const getGroupByUserId = async (req, res, next) => {
   try {
     const { userId } = req.params;
     const channels= await groupService.getGroupsByUser(userId);
-    res.json(channels);
+    sendServiceResult(res, channels);
   } catch (error) {
     next(error);
   }
@@ -23,8 +28,8 @@ const getGroupByUserId = async (req, res, next) => {
 
 const createGroup = async (req, res, next) => {
   try {
-    const result = await groupService.createGroup(req.body);
-    res.status(201).json(result);
+    const result = await groupService.createGroup({ ...req.body, userId: req.user.id });
+    sendServiceResult(res, result, 201);
   } catch (error) {
     next(error);
   }
@@ -33,8 +38,8 @@ const createGroup = async (req, res, next) => {
 const updateGroup = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const result = await groupService.updateGroup({ id, group: req.body });
-    res.json(result);
+    const result = await groupService.updateGroup({ id, group: req.body, userId: req.user.id });
+    sendServiceResult(res, result);
   } catch (error) {
     next(error);
   }
@@ -43,8 +48,8 @@ const updateGroup = async (req, res, next) => {
 const deleteGroup = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const result = await groupService.deleteGroup(id);
-    res.json(result);
+    const result = await groupService.deleteGroup({ id, userId: req.user.id });
+    sendServiceResult(res, result);
   } catch (error) {
     next(error);
   }
